Separating dataverse types and functions from logic.go
