--- manifests/subtree-whitelist/demo.mramsey
@@ -1,8 +1,0 @@
-name=Mack Ramsey Dataverse
-identifier=mramsey
-url=https://demo.dataverse.org/dataverse/mramsey
-image_url=
-description=Test dataverse for training purposes.
-published_at=2015-04-20T10:14:18Z
-server_id=demo_dataverse
-server_url=https://demo.dataverse.org
