--- manifests/subtree-whitelist/demo.splash
@@ -1,8 +1,0 @@
-name=Splish Splash Dataverse
-identifier=splash
-url=https://demo.dataverse.org/dataverse/splash
-image_url=
-description=
-published_at=2015-04-20T14:40:11Z
-server_id=demo_dataverse
-server_url=https://demo.dataverse.org
