--- manifests/subtree-whitelist/demo.test
@@ -1,8 +1,0 @@
-name=Investigation of Test Dataverse
-identifier=test
-url=https://demo.dataverse.org/dataverse/test
-image_url=
-description=test investigation
-published_at=2015-04-20T14:11:51Z
-server_id=demo_dataverse
-server_url=https://demo.dataverse.org
